+++ api/models/challenge_model.js
@@ -5,7 +5,7 @@ const testcaseSchema = new mongoose.Schema({
   expected_output: { type: String, required: true },
 });
 
-const problemDescriptionSchema = new mongoose.Schema({
+const challengeDescriptionSchema = new mongoose.Schema({
   description: { type: String, required: true },
   input_format: { type: String, required: true },
   output_format: { type: String, required: true },
@@ -19,22 +19,20 @@ const skeletonCodeSchema = new mongoose.Schema({
 
 const challengeSchema = new mongoose.Schema(
   {
-    challenge_title: { type: String, required: true }, // Renamed from question_name
-    problem_description: { type: problemDescriptionSchema, required: true },
-    question_difficulty: {
+    challenge_title: { type: String, required: true }, // Renamed from challenge_name
+    challenge_description: { type: challengeDescriptionSchema, required: true },
+    challenge_difficulty: {
       type: Number,
       required: true,
       min: 1,
       max: 10,
     },
-    question_type: { type: String, required: true },
     users_attempted: { type: Number, default: 0 },
     users_solved: { type: Number, default: 0 },
     languages_supported: { type: [String], required: true },
-    daily_question: { type: Date },
-    subject: { type: String, required: true },
-    test_cases: [testcaseSchema], // Updated from question_testcases
-    question_solutions: { type: [String], required: false },
+    daily_challenge: { type: Date },
+    test_cases: [testcaseSchema], // Updated from challenge_testcases
+    challenge_solutions: { type: [String], required: false },
     skeleton_code: skeletonCodeSchema, // New field for skeleton code
     hints: { type: [String], default: [] }, // New field for hints
     subject_tags: { type: [String], default: [], required: false }, // New field for subject_tags
