+++ app.js
@@ -4,7 +4,7 @@ import dotenv from 'dotenv';  // Import dotenv for environment variables
 import userRoutes from './api/routes/user_routes.js';  // Import user routes (note the .js extension)
 import submissionRoutes from './api/routes/submission_routes.js';  // Import submission routes
 import subjectRoutes from './api/routes/subject_routes.js';  // Import subject routes
-import problemRoutes from './api/routes/problem_routes.js';  // Import problem routes
+import challengeRoutes from './api/routes/challenge_routes.js';  // Import problem routes
 import cors from 'cors';
 
 // Initialize express app
@@ -26,7 +26,7 @@ mongoose.connect(process.env.MONGO_URI)
 app.use('/api/users', userRoutes);  // Prefix routes with /api/users
 app.use('/api/submissions', submissionRoutes);  // Prefix routes with /api/submissions
 app.use('/api/subjects', subjectRoutes);  // Prefix routes with /api/subjects
-app.use('/api/problems', problemRoutes);  // Prefix routes with /api/problems
+app.use('/api/challenges', challengeRoutes);  // Prefix routes with /api/problems
 
 // Start the server
 const PORT = process.env.PORT || 3000;
